+++ src/collective/documentgenerator/tests/test_base_helper_view.py
@@ -58,6 +58,13 @@ class TestBaseHelperViewMethods(DexterityIntegrationTests):
         # unknown variable
         self.assertEqual(helper_view.context_var('dexter', 'undefined'), 'undefined')
 
+        # None value (after edition, an empty value is replaced by None)
+        pod_template.context_variables = [{'name': u'dexter', 'value': None}]
+        generation_context = view._get_generation_context(helper_view, pod_template=pod_template)
+        renderer = appy.pod.renderer.Renderer(StringIO(document_template.data), generation_context, 'dummy.odt')
+        helper_view._set_appy_renderer(renderer)
+        self.assertEqual(helper_view.context_var('dexter'), '')  # default is '' by default
+
         # add a context variable on pod template
         pod_template.context_variables = [{'name': u'dexter', 'value': u'1'}]
         generation_context = view._get_generation_context(helper_view, pod_template=pod_template)